Rename test module to pnettest
